# src/main/webapp/docs/swagger.yaml
openapi: 3.0.0
servers: 
  - url: 'https://api-electronicdevices-313314.nw.r.appspot.com/api/v1/'
  - url: 'http://localhost:8089/api/v1/'
info:
  description: This API can be used to consult multiple electronic devices which could be found in different shops.
  version: 2.0.0
  title: API-ElectronicDevices
  contact:
    email: joselugarciamarin2406@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /complements:
    get:
      tags:
        - Complements
      summary: Get complements
      description: Get all the complements
      operationId: getComplements
      parameters:
        - name: order
          in: query
          description: Order results by name or price
          required: false
          schema:
            type: string
            enum:
              - name
              - '-name'
              - price
              - '-price'
        - name: isEmpty
          in: query
          description: Filter complements with no devices
          required: false
          schema:
            type: boolean
        - name: name
          in: query
          description: Filter complements by name
          required: false
          schema: 
            type: string
        - name: material
          in: query
          description: Filter complements by material
          schema:
            type: string
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complement'
        '400':
          description: Bad request
    post:
      tags:
        - Complements
      summary: Create a complement
      description: Create a complement
      operationId: createComplement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplementEmpty'
        description: Device to create
        required: true
      responses:
        '201':
          description: Complement created
        '400':
          description: Bad reeques
    put:
      tags:
        - Complements
      summary: Update a complement
      description: Upadate a complement
      operationId: updateComplement
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Complement'
        description: Complement to update
        required: true
      responses:
        '204':
          description: Complement updated
        '400':
          description: Bad request
  '/complements/{id}':
    get:
      tags:
        - Complements
      summary: Get a complement by id
      description: Get a complement  by id
      operationId: getComplement
      parameters:
        - name: id
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses: 
        '200':
          description: Complement found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Complement'
        '400':
          description: Complement not found
    delete:
      tags:
        - Complements
      summary: Delete a complement
      description: Delete a complement
      operationId: deleteComplement
      parameters:
        - name: id
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Complement deleted
        '404':
          description: Complement not found
  '/complements/{complementId}/{deviceId}':
    post:
      tags: 
        - Complements
      summary: Add a device compatible with a complemet
      description: Add a device compatible with a complement
      operationId: addDeviceToComplement
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      parameters: 
        - name: complementId
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: deviceId
          in: path
          description: Id of the device
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '201':
          description: Device added to the compatible devices list with a complement
        '400':
          description: Bad request
        '404':
          description: Accessory or device not found
    delete:
      tags: 
        - Complements
      summary: Delete a device compatible with a complement
      description: Delete a device compatible with a complement
      operationId: deleteDeviceFromComplement
      parameters: 
        - name: complementId
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: deviceId
          in: path
          description: Id of the device
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Device added to the compatible devices list with a complement 
        '404':
          description: Accessory or device not found
...
  '/shops/complements/{shopId}/{complementId}':
    post:
      tags: 
        - Shops
      summary: Add complement to shop
      description: Add complement to shop
      operationId: addComplementToShop
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      parameters: 
        - name: shopId
          in: path
          description: Id of the shop
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: complementId
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '201':
          description: Complement added to the shop
        '400':
          description: Bad request
        '404':
          description: Shop or complement not found
    delete:
      tags: 
        - Shops
      summary: Delete a complement from a shop
      description: Delete a complement from a shop
      operationId: deleteComplementFromShop
      parameters: 
        - name: shopId
          in: path
          description: Id of the shop
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: complementId
          in: path
          description: Id of the complement
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
      responses:
        '204':
          description: Complement delete from shop
        '404':
          description: Shop or complement not found
...
components:
  schemas:
...
    Complement:
      type: array
      items: 
        type: object
        required:
          - name
        properties:
          complementId:
            type: integer
          name: 
            type: string
          material: 
            type: string
          price: 
            type: number
          compatibleDevices:
            type: array
            items: 
              type: object
              required: 
                - name
              properties:
                id:
                  type: integer
                name:
                  type: string
                brand:
                  type: string
                price: 
                  type: number
                screenSize:
                  type: number
                weight:
                  type: number
                deviceType:
                  type: string
                storage:
                  type: integer
                processor:
                  type: string
                memory:
                  type: integer
                OS:
                  type: string
                setColors:
                  type: array
                  items:
                    type: string
                camera:
                  type: string
      example:
        complementId: 25
        name: Silicon Case Iphone 12 Pro Max
        material: Silicon
        price: 50.0
        compatibleDevice:
          id: 1
          name: Galaxy s20 FE
          brand: Samsung
          price: 649.0
          screenSize: 6.5
          weight: 0.19
          deviceType: MOBILE
          storage: 128
          processor: Samsung Exynos 990
          memory: 6
          setColors: [Cloud Orange, Cloud Red, Cloud white, Cloud Navy]
          camera: "Principal 12Mp"
          os: Android
    ComplementEmpty:
      type: array
      items: 
        type: object
        required:
          - name
        properties:
          complementId:
            type: integer
          name: 
            type: string
          material: 
            type: string
          pirce:
            type: number
      example:
        complementId: 25
        name: Silicon Case Iphone 12 Pro Max
        material: Silicon
        price: 50.0
        compatibleDevice:
          id: 1
          name: Galaxy s20 FE
          brand: Samsung
          price: 649.0
          screenSize: 6.5
          weight: 0.19
          deviceType: MOBILE
          storage: 128
          processor: Samsung Exynos 990
          memory: 6
          setColors: [Cloud Orange, Cloud Red, Cloud white, Cloud Navy]
          camera: "Principal 12Mp"
          os: Android
...